[shared libs] Break http and dd_send apart
These were shipped as single-header libraries in kind of a stupid way.
